requirements.txt
twscrape>=0.6.0
scrapy>=2.11.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
pandas>=1.5.0
bcrypt>=4.0.0
tweepy>=4.14.0